# src/main/resources/openapi/yaml/api.yaml
openapi: 3.0.0
info:
  title: User Registration API
  version: 1.0.0
  description: |-
    Job Beacon User Registration API
  termsOfService: 'https://www.jobbeacon.com/terms'
  contact:
    name: Job Beacon Support
    url: 'https://www.jobbeacon.com/support'
    email: andrey.kandzuba@gmail.com
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: '{protocol}:/api/v1'
    variables:
      protocol:
        enum:
          - http
          - https
        default: https
paths:
  /users:
    post:
      summary: Register a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /users/{username}:
    get:
      summary: Retrieve a user by username
      operationId: findUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    put:
      summary: Update a user
      operationId: updateUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '404':
          description: User not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a user
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found
        '500':
          description: Internal server error
components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
          example: lopez.paulette123
        email:
          type: string
          format: email
          example: lopez.paulette123@mail.com
        firstName:
          type: string
          example: Lopez
        lastName:
          type: string
          example: Paulette
        phone:
          type: string
          example: 123-456-7890
        street:
          type: string
          example: '123 Main St'
        city:
          type: string
          example: New York
        state:
          type: string
          example: NY
        zip:
          type: string
          example: 10001
        country:
          type: string
          example: USA
      required:
        - username
        - email
        - firstName
        - lastName
        - phone
        - street
        - city
        - state
        - zip
        - country